change api to storage and deleted errors components and logic for errors
