--- src/types/ErrorEnum.ts
@@ -1,7 +1,0 @@
-export enum ErrorEnum {
-  LOAD = 'Unable to load todos',
-  TITLE = 'Title should not be empty',
-  ADD = 'Unable to add a todo',
-  DELETE = 'Unable to delete a todo',
-  UPDATE = 'Unable to update a todo',
-}
