+++ src/types/Todo.ts
@@ -1,6 +1,5 @@
 export interface Todo {
   id: number;
-  userId: number;
   title: string;
   completed: boolean;
 }
